DROP column refinements
Updates as per code review.
